requirements.txt
numpy 
pandas
scilkit-learn
matplotlib
seaborn
streamlit
transformers
Image
Pillow